fix: correct triggered_by_user_id column type to match users.user_id

+++ database/migrations/2026_07_27_000001_create_backups_table.php
@@ -1,9 +1,7 @@
 <?php
-
 use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
-
 return new class extends Migration
 {
     public function up()
@@ -11,20 +9,18 @@ return new class extends Migration
         Schema::create('backups', function (Blueprint $table) {
             $table->increments('backup_id');
             $table->string('trigger_type');
-            $table->unsignedInteger('triggered_by_user_id')->nullable();
+            $table->unsignedBigInteger('triggered_by_user_id')->nullable();
             $table->string('file_path')->nullable();
             $table->unsignedBigInteger('file_size')->nullable();
             $table->string('status')->default('pending');
             $table->text('error_message')->nullable();
             $table->timestamps();
-
             $table->foreign('triggered_by_user_id')
                 ->references('user_id')
                 ->on('users')
                 ->nullOnDelete();
         });
     }
-
     public function down()
     {
         Schema::dropIfExists('backups');
